修正progress计算错误

+++ turquoise.imageloader/src/main/java/sviolet/turquoise/x/imageloader/server/DiskCacheServer.java
@@ -345,7 +345,7 @@ public class DiskCacheServer extends DiskCacheModule {
         float progress = -1;
         //check progress
         if (task.getLoadProgress().total() > 0) {
-            progress = task.getLoadProgress().loaded() / task.getLoadProgress().total();
+            progress = (float)task.getLoadProgress().loaded() / (float)task.getLoadProgress().total();
             if (progress > getComponentManager().getServerSettings().getAbortOnLowNetworkSpeedBoundaryProgress()){
                 return false;
             }
